♻️ 유저 프로필 이미지 부분 파일 분리 (#29)
