+++ src/components/userSettings/UserProfile.tsx
@@ -1,26 +1,10 @@
-import { SyntheticEvent } from 'react';
 import styled from 'styled-components';
-import Background from '@src/assets/images/userSettings/background_default.svg';
-import Profile from '@src/assets/images/userSettings/profile_default.svg';
+import UserProfilImg from '@src/components/userSettings/UserProfileImg';
 
 const UserProfile = () => {
-  const handleImgError = (
-    e: SyntheticEvent<HTMLImageElement>,
-    type: string,
-  ) => {
-    if (type === 'back') {
-      e.currentTarget.src = Background;
-      return;
-    }
-    e.currentTarget.src = Profile;
-  };
-
   return (
     <SLayout>
-      <SImages>
-        <SBackgroundImg src='url' onError={(e) => handleImgError(e, 'back')} />
-        <SProfileImg src='url' onError={(e) => handleImgError(e, 'profile')} />
-      </SImages>
+      <UserProfilImg />
       <SContainer>
         <SNickname>내별명</SNickname>
         <SMountain>0번째, ㅇㅇ산 등산가</SMountain>
@@ -57,30 +41,6 @@ const SLayout = styled.div`
   background-color: ${({ theme }) => theme.colors.white};
 
   text-align: center;
-`;
-const SImages = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  width: 100%;
-`;
-const SBackgroundImg = styled.img`
-  width: 100%;
-  height: 10.4688rem;
-
-  border-radius: 1.875rem 1.875rem 0 0;
-
-  object-fit: cover;
-`;
-const SProfileImg = styled.img`
-  margin-top: -3.125rem;
-  width: 6.25rem;
-  height: 6.25rem;
-
-  border-radius: 50%;
-
-  object-fit: contain;
 `;
 const SContainer = styled.div`
   display: flex;
